requirements.txt
requests>=2.2.1
pyopenssl>=0.13
asn1crypto>=0.15.1
